requirements.txt
streamlit
pandas
numpy
requests
matplotlib
python-dotenv
beautifulsoup4
openai
